Refacter owner's resource Related to issue #4 Related to issue #2
