+++ src/Categories/Owners/index.js
@@ -1,0 +1,4 @@
+export { OwnerList } from './OwnerList';
+export { OwnerShow } from './OwnerShow';
+export { OwnerCreate } from './OwnerCreate';
+export { OwnerEdit } from './OwnerEdit';
